refactor: move cli-only utils out of staged-sync
