+++ bin/reth/src/chain/init.rs
@@ -1,10 +1,10 @@
-use crate::dirs::{DataDirPath, MaybePlatformPath};
+use crate::{
+    dirs::{DataDirPath, MaybePlatformPath},
+    utils::genesis_value_parser,
+};
 use clap::Parser;
 use reth_primitives::ChainSpec;
-use reth_staged_sync::utils::{
-    chainspec::genesis_value_parser,
-    init::{init_db, init_genesis},
-};
+use reth_staged_sync::utils::init::{init_db, init_genesis};
 use std::sync::Arc;
 use tracing::info;
