+++ bin/reth/src/debug_cmd/execution.rs
@@ -4,7 +4,7 @@ use crate::{
     dirs::{DataDirPath, MaybePlatformPath},
     node::events,
     runner::CliContext,
-    utils::get_single_header,
+    utils::{genesis_value_parser, get_single_header},
 };
 use clap::Parser;
 use futures::{stream::select as stream_select, StreamExt};
@@ -27,10 +27,8 @@ use reth_network::NetworkHandle;
 use reth_network_api::NetworkInfo;
 use reth_primitives::{stage::StageId, BlockHashOrNumber, BlockNumber, ChainSpec, H256};
 use reth_provider::{providers::get_stage_checkpoint, ShareableDatabase, Transaction};
-use reth_staged_sync::utils::{
-    chainspec::genesis_value_parser,
-    init::{init_db, init_genesis},
-};
+use reth_staged_sync::utils::init::{init_db, init_genesis};
+
 use reth_stages::{
     sets::DefaultStages,
     stages::{
