+++ bin/reth/src/debug_cmd/merkle.rs
@@ -1,5 +1,8 @@
 //! Command for debugging merkle trie calculation.
-use crate::dirs::{DataDirPath, MaybePlatformPath};
+use crate::{
+    dirs::{DataDirPath, MaybePlatformPath},
+    utils::genesis_value_parser,
+};
 use clap::Parser;
 use reth_db::{cursor::DbCursorRO, tables, transaction::DbTx};
 use reth_primitives::{
@@ -7,7 +10,7 @@ use reth_primitives::{
     ChainSpec,
 };
 use reth_provider::Transaction;
-use reth_staged_sync::utils::{chainspec::genesis_value_parser, init::init_db};
+use reth_staged_sync::utils::init::init_db;
 use reth_stages::{
     stages::{
         AccountHashingStage, ExecutionStage, ExecutionStageThresholds, MerkleStage,
