+++ bin/reth/src/p2p/mod.rs
@@ -2,7 +2,7 @@
 use crate::{
     args::{get_secret_key, DiscoveryArgs},
     dirs::{DataDirPath, MaybePlatformPath},
-    utils::get_single_header,
+    utils::{chain_spec_value_parser, get_single_header, hash_or_num_value_parser},
 };
 use backon::{ConstantBuilder, Retryable};
 use clap::{Parser, Subcommand};
@@ -12,7 +12,6 @@ use reth_discv4::NatResolver;
 use reth_interfaces::p2p::bodies::client::BodiesClient;
 use reth_primitives::{BlockHashOrNumber, ChainSpec, NodeRecord};
 use reth_provider::ShareableDatabase;
-use reth_staged_sync::utils::{chainspec::chain_spec_value_parser, hash_or_num_value_parser};
 use std::{path::PathBuf, sync::Arc};
 
 /// `reth p2p` command
