+++ crates/staged-sync/src/utils/mod.rs
@@ -1,114 +1,4 @@
 //! Utility functions.
-use reth_primitives::{BlockHashOrNumber, H256};
-use std::{
-    env::VarError,
-    net::{IpAddr, Ipv4Addr, SocketAddr, ToSocketAddrs},
-    path::{Path, PathBuf},
-    str::FromStr,
-};
-use walkdir::{DirEntry, WalkDir};
-
-/// Utilities for parsing chainspecs
-pub mod chainspec;
 
 /// Utilities for initializing parts of the chain
 pub mod init;
-
-/// Finds all files in a directory with a given postfix.
-pub fn find_all_files_with_postfix(path: &Path, postfix: &str) -> Vec<PathBuf> {
-    WalkDir::new(path)
-        .into_iter()
-        .filter_map(|e| e.ok())
-        .filter(|e| e.file_name().to_string_lossy().ends_with(postfix))
-        .map(DirEntry::into_path)
-        .collect::<Vec<PathBuf>>()
-}
-
-/// Parses a user-specified path with support for environment variables and common shorthands (e.g.
-/// ~ for the user's home directory).
-pub fn parse_path(value: &str) -> Result<PathBuf, shellexpand::LookupError<VarError>> {
-    shellexpand::full(value).map(|path| PathBuf::from(path.into_owned()))
-}
-
-/// Parse [BlockHashOrNumber]
-pub fn hash_or_num_value_parser(value: &str) -> Result<BlockHashOrNumber, eyre::Error> {
-    match H256::from_str(value) {
-        Ok(hash) => Ok(BlockHashOrNumber::Hash(hash)),
-        Err(_) => Ok(BlockHashOrNumber::Number(value.parse()?)),
-    }
-}
-
-/// Error thrown while parsing a socket address.
-#[derive(thiserror::Error, Debug)]
-pub enum SocketAddressParsingError {
-    /// Failed to convert the string into a socket addr
-    #[error("Cannot parse socket address: {0}")]
-    Io(#[from] std::io::Error),
-    /// Input must not be empty
-    #[error("Cannot parse socket address from empty string")]
-    Empty,
-    /// Failed to parse the address
-    #[error("Could not parse socket address from {0}")]
-    Parse(String),
-    /// Failed to parse port
-    #[error("Could not parse port: {0}")]
-    Port(#[from] std::num::ParseIntError),
-}
-
-/// Parse a [SocketAddr] from a `str`.
-///
-/// The following formats are checked:
-///
-/// - If the value can be parsed as a `u16` or starts with `:` it is considered a port, and the
-/// hostname is set to `localhost`.
-/// - If the value contains `:` it is assumed to be the format `<host>:<port>`
-/// - Otherwise it is assumed to be a hostname
-///
-/// An error is returned if the value is empty.
-pub fn parse_socket_address(value: &str) -> Result<SocketAddr, SocketAddressParsingError> {
-    if value.is_empty() {
-        return Err(SocketAddressParsingError::Empty)
-    }
-
-    if let Some(port) = value.strip_prefix(':').or_else(|| value.strip_prefix("localhost:")) {
-        let port: u16 = port.parse()?;
-        return Ok(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), port))
-    }
-    if let Ok(port) = value.parse::<u16>() {
-        return Ok(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), port))
-    }
-    value
-        .to_socket_addrs()?
-        .next()
-        .ok_or_else(|| SocketAddressParsingError::Parse(value.to_string()))
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use rand::{thread_rng, Rng};
-
-    #[test]
-    fn parse_socket_addresses() {
-        for value in ["localhost:9000", ":9000", "9000"] {
-            let socket_addr = parse_socket_address(value)
-                .unwrap_or_else(|_| panic!("could not parse socket address: {value}"));
-
-            assert!(socket_addr.ip().is_loopback());
-            assert_eq!(socket_addr.port(), 9000);
-        }
-    }
-
-    #[test]
-    fn parse_socket_address_random() {
-        let port: u16 = thread_rng().gen();
-
-        for value in [format!("localhost:{port}"), format!(":{port}"), port.to_string()] {
-            let socket_addr = parse_socket_address(&value)
-                .unwrap_or_else(|_| panic!("could not parse socket address: {value}"));
-
-            assert!(socket_addr.ip().is_loopback());
-            assert_eq!(socket_addr.port(), port);
-        }
-    }
-}
